requirements.txt
Django==5.2.1
djangorestframework==3.16.0
drf-yasg>=1.21.10
Pillow==11.2.1
psycopg2-binary==2.9.10
gunicorn==22.0.0  # Для production
whitenoise==6.6.0  # Для статики
python-dotenv==1.0.1  # Для переменных окружения
django-filter==25.1